Introduce theme colors, remove borders, add box shadow
- Theme colors were added in order to enhance display.
- Theme colors were stored in CSS variables in order to allow
easy color swapping.
- Removed borders from page elements, since element positioning
can be seen through color differentiation.
- Updated remaining borders to be colored in accordance with
theme colors for better visual consistency.
- Added box shadow to elements to create depth.
- A box shadow variable was created under :root in order to ensure
box shadow consistency. This is important because box shadow
consistency gives the illusion that all box shadows were created
from the same light source, which is important in creating a
realistic depth effect.
- Multiple box shadows were given to each element to promote a more
natural looking shadow than what the box-shadow property can
provide with a single shadow alone.
- Added a --box-shadow-elevated variable which is a lightened
version of the --box-shadow variable that can be used to support
differing levels of depth on the page.
- Added text shadow to <nav> text when hovered to support the
illusion that it is moving out of the page when hovered.